requirements.txt
# Slack Lists MCP Server Dependencies

# MCP (Model Context Protocol) SDK
mcp>=1.2.0

# HTTP client for Slack API requests
httpx>=0.25.0

# Environment variable loading
python-dotenv>=1.0.0

# SSE transport dependencies (for remote MCP)
starlette>=0.27.0
uvicorn>=0.23.0
sse-starlette>=1.6.0

# Standard library dependencies (included for completeness)
# These are typically included with Python but listed for clarity
typing-extensions>=4.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: For enhanced logging and monitoring
structlog>=23.0.0

